requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
python-multipart
celery
pillow
redis
rabbitmq-client  # not required, just run rabbitmq server; celery depends on kombu
